Move `Pointer` to its own module
